Refactoring changes: split up the downloaders utilities
