+++ src/input_data/downloaders/__init__.py
@@ -1,0 +1,5 @@
+"""
+Dataset utilities for downloading and extracting common ML datasets.
+"""
+
+from .datasets import DownloadInfo, download_and_extract_archive
